refactor(UserDashboard): extract tabs into modular components
